+++ src/utils/spell.cairo
@@ -9,7 +9,7 @@ mod spell_utils {
 
     fn spell_effect(
         world: IWorldDispatcher,
-        entity_id: u16,
+        entity_id: u8,
         card_id: u16,
         ref battle: Battle,
         ref monster: Monster,
